updating the similarity calculation script

+++ calculateSimilarity.sh
@@ -15,6 +15,7 @@ DIR="$BASE/timepoints/*"
 count=0
 echo "Timepoint, Similarity, Mutual_Information" > "$BASE/similarities_preregistration.csv"
 for img in $DIR ; do
+    # make sure $img != 'template.nii.gz'
     # compare each image to the template
     sims=$(./utils/similarity.sh $TEMPLATE $img)
     echo $count, $sims >> "$BASE/similarities_preregistration.csv"
@@ -22,23 +23,27 @@ for img in $DIR ; do
 done
 
 DIR="$BASE/markov/*"
-# for all the images in the non-registered version
-count=0
-echo "Timepoint, Similarity, Mutual_Information" > "$BASE/similarities_markov.csv"
-for img in $DIR ; do
-    # compare each image to the template
-    sims=$(./utils/similarity.sh $TEMPLATE $img)
-    echo $count, $sims >> "$BASE/similarities_markov.csv"
-    count=$((count+1))
-done
+if [ -d $DIR ] # check that the directory exists
+    # for all the images in the non-registered version
+    count=0
+    echo "Timepoint, Similarity, Mutual_Information" > "$BASE/similarities_markov.csv"
+    for img in $DIR ; do
+        # compare each image to the template
+        sims=$(./utils/similarity.sh $TEMPLATE $img)
+        echo $count, $sims >> "$BASE/similarities_markov.csv"
+        count=$((count+1))
+    done
+fi
 
 DIR="$BASE/bifur-markov/*"
-# for all the images in the non-registered version
-count=0
-echo "Timepoint, Similarity, Mutual_Information" > "$BASE/similarities_bifur_markov.csv"
-for img in $DIR ; do
-    # compare each image to the template
-    sims=$(./utils/similarity.sh $TEMPLATE $img)
-    echo $count, $sims >> "$BASE/similarities_bifur_markov.csv"
-    count=$((count+1))
-done
+# for all the images in the bifurcating-markov version
+if [ -d $DIR ] # check that the directory exists
+    count=0
+    echo "Timepoint, Similarity, Mutual_Information" > "$BASE/similarities_bifur_markov.csv"
+    for img in $DIR ; do
+        # compare each image to the template
+        sims=$(./utils/similarity.sh $TEMPLATE $img)
+        echo $count, $sims >> "$BASE/similarities_bifur_markov.csv"
+        count=$((count+1))
+    done
+fi
